feat: tweaks, cli is already externalized
