+++ .circleci/config.yml
@@ -55,7 +55,7 @@ jobtest: &jobtest
         name: Installing project dependencies
         command: yarn install --prefer-offline || yarn install
     - run:
-        name: Testing your project
+        name: Linting and testing your project
         command: yarn global add nyc@latest && yarn run lint && yarn test
     - save_cache:
         key: release-{{ checksum "yarn.lock" }}
@@ -97,7 +97,7 @@ jobs:
   #   <<: *node10osx
   #   <<: *jobtest
 
-  release:
+  publish:
     <<: *node10linux
     steps:
       - checkout
@@ -107,7 +107,7 @@ jobs:
           command: yarn run build || echo "No build step."
       - run:
           name: Releasing and publishing
-          command: yarn run release
+          command: yarn run release || (yarn global add @tunnckocore/release-cli && yarn run release)
 
 workflows:
   version: 2
@@ -137,7 +137,7 @@ workflows:
 
       # Release and NPM publish
       # Executed only on master
-      - release:
+      - publish:
           requires:
             # - test-node8osx
             # - test-node10osx
