+++ .nycrc.json
@@ -1,16 +1,3 @@
 {
-  "statements": 0,
-  "branches": 0,
-  "functions": 0,
-  "lines": 0,
-  "cache": true,
-  "check-coverage": true,
-  "require": "esm",
-  "reporter": [
-    "lcov",
-    "text"
-  ],
-  "include": [
-    "src/**/*.js"
-  ]
+  "extends": "./node_modules/@tunnckocore/config/nyc.json"
 }
