+++ README.md
@@ -1,6 +1,6 @@
-# {%= name %} [![npm version][npmv-img]][npmv-url] [![github release][ghrelease-img]][ghrelease-url] [![License][license-img]][license-url]
+# @tunnckocore/release [![npm version][npmv-img]][npmv-url] [![github release][ghrelease-img]][ghrelease-url] [![License][license-img]][license-url]
 
-> {%= description %}
+> Release and publish a new project to NPM and GitHub, automatically. Following Conventional Commits specification and Semver. It provides GitHub App, similar to [semantic-release][], too.
 
 Please consider following this project's author, [Charlike Mike Reagent](https://github.com/tunnckoCore), and :star: the project to show your :heart: and support.
 
@@ -39,11 +39,20 @@ Project is [semantically](https://semver.org) & automatically released on [Circl
 
 ## Table of Contents
 
-<!-- toc -->
+- [Install](#install)
+- [See Also](#see-also)
+- [Contributing](#contributing)
+  * [Follow the Guidelines](#follow-the-guidelines)
+  * [Support the project](#support-the-project)
+  * [OPEN Open Source](#open-open-source)
+  * [Wonderful Contributors](#wonderful-contributors)
+- [License](#license)
+
+_(TOC generated by [verb](https://github.com/verbose/verb) using [markdown-toc](https://github.com/jonschlinkert/markdown-toc))_
 
 ## Install
 
-This project requires [**Node.js**](https://nodejs.org) **{%= engines.node %}**. Install it using
+This project requires [**Node.js**](https://nodejs.org) **^8.11.0 || >=10.13.0**. Install it using
 [**yarn**](https://yarnpkg.com) or [**npm**](https://npmjs.com).  
 _We highly recommend to use Yarn when you think to contribute to this project._
 
@@ -55,17 +64,18 @@ For NPM publishing on the CI or locally, install [@tunnckocore/release-cli](http
 
 **[back to top](#thetop)**
 
-{% if (verb.related && verb.related.list && verb.related.list.length) { %}
-
 ## See Also
 
 Some of these projects are used here or were inspiration for this one, others are just related. So, thanks for your existance!
 
-{%= related(verb.related.list, { words: 10 }) %}
+- [asia](https://www.npmjs.com/package/asia): Blazingly fast, magical and minimalist testing framework, for Today and… [more](https://github.com/olstenlarck/asia#readme) | [homepage](https://github.com/olstenlarck/asia#readme "Blazingly fast, magical and minimalist testing framework, for Today and Tomorrow")
+- [detect-next-version](https://www.npmjs.com/package/detect-next-version): Calculates next version, based on given commit message and following… [more](https://github.com/tunnckoCoreLabs/detect-next-version) | [homepage](https://github.com/tunnckoCoreLabs/detect-next-version "Calculates next version, based on given commit message and following Conventional Commits")
+- [docks](https://www.npmjs.com/package/docks): Extensible system for parsing and generating documentation. It just freaking… [more](https://github.com/tunnckoCore/docks) | [homepage](https://github.com/tunnckoCore/docks "Extensible system for parsing and generating documentation. It just freaking works!")
+- [git-commits-since](https://www.npmjs.com/package/git-commits-since): Get all commits since given period of time or by… [more](https://github.com/tunnckoCoreLabs/git-commits-since) | [homepage](https://github.com/tunnckoCoreLabs/git-commits-since "Get all commits since given period of time or by default from latest git semver tag. Understands and follows both SemVer and the Conventional Commits specification.")
+- [gitcommit](https://www.npmjs.com/package/gitcommit): Lightweight and joyful `git commit` replacement. Conventional Commits compliant. | [homepage](https://github.com/tunnckoCore/gitcommit "Lightweight and joyful `git commit` replacement. Conventional Commits compliant.")
+- [parse-commit-message](https://www.npmjs.com/package/parse-commit-message): Extensible parser for git commit messages following Conventional Commits Specification | [homepage](https://github.com/tunnckoCoreLabs/parse-commit-message "Extensible parser for git commit messages following Conventional Commits Specification")
 
 **[back to top](#thetop)**
-
-{% } %}
 
 ## Contributing
 
@@ -77,7 +87,7 @@ For bugs reports and feature requests, [please create an issue][open-issue-url] 
 
 ### Support the project
 
-[Become a Partner or Sponsor?][patreon-url] :dollar: Check the **Partner**, **Sponsor** or **Omega-level** tiers! :tada: You can get your company logo, link & name on this file. It's also rendered on package page in [npmjs.com][npmv-url] and [yarnpkg.com](https://yarnpkg.com/en/package/{%= name %}) sites too! :rocket:
+[Become a Partner or Sponsor?][patreon-url] :dollar: Check the **Partner**, **Sponsor** or **Omega-level** tiers! :tada: You can get your company logo, link & name on this file. It's also rendered on package page in [npmjs.com][npmv-url] and [yarnpkg.com](https://yarnpkg.com/en/package/@tunnckocore/release) sites too! :rocket:
 
 Not financial support? Okey! [Pull requests](https://github.com/tunnckoCoreLabs/contributing#opening-a-pull-request), stars and all kind of [contributions](https://opensource.guide/how-to-contribute/#what-it-means-to-contribute) are always
 welcome. :sparkles:
@@ -119,14 +129,14 @@ Released under the [Apache-2.0 License][license-url].
 
 <!-- Heading badges -->
 
-[npmv-url]: https://www.npmjs.com/package/{%= name %}
-[npmv-img]: https://badgen.net/npm/v/{%= name %}?icon=npm
+[npmv-url]: https://www.npmjs.com/package/@tunnckocore/release
+[npmv-img]: https://badgen.net/npm/v/@tunnckocore/release?icon=npm
 
 [ghrelease-url]: https://github.com/tunnckoCoreLabs/release/releases/latest
 [ghrelease-img]: https://badgen.net/github/release/tunnckoCoreLabs/release?icon=github
 
 [license-url]: https://github.com/tunnckoCoreLabs/release/blob/master/LICENSE
-[license-img]: https://badgen.net/npm/license/{%= name %}
+[license-img]: https://badgen.net/npm/license/@tunnckocore/release
 
 <!-- Front line badges -->
 
@@ -147,9 +157,9 @@ Released under the [Apache-2.0 License][license-url].
 [new-release-url]: https://ghub.io/new-release
 [new-release-img]: https://badgen.net/badge/semantically/released/05c5ff
 
-[downloads-weekly-img]: https://badgen.net/npm/dw/{%= name %}
-[downloads-monthly-img]: https://badgen.net/npm/dm/{%= name %}
-[downloads-total-img]: https://badgen.net/npm/dt/{%= name %}
+[downloads-weekly-img]: https://badgen.net/npm/dw/@tunnckocore/release
+[downloads-monthly-img]: https://badgen.net/npm/dm/@tunnckocore/release
+[downloads-total-img]: https://badgen.net/npm/dt/@tunnckocore/release
 
 [renovateapp-url]: https://renovatebot.com
 [renovateapp-img]: https://badgen.net/badge/renovate/enabled/green
@@ -164,3 +174,6 @@ Released under the [Apache-2.0 License][license-url].
 [shareu]: https://twitter.com/intent/tweet?text=https://github.com/tunnckoCoreLabs/release&via=tunnckoCore
 [shareb]: https://badgen.net/badge/twitter/share/1da1f2?icon=twitter
 [open-issue-url]: https://github.com/tunnckoCoreLabs/release/issues/new
+
+[new-release]: https://github.com/tunnckoCore/new-release
+[semantic-release]: https://github.com/semantic-release/semantic-release
