--- now.json
@@ -1,10 +1,0 @@
-{
-  "public": true,
-  "alias": "tunnckocore-release",
-  "env": {
-    "APP_ID": "@tunnckocore-release-id",
-    "NODE_ENV": "production",
-    "PRIVATE_KEY": "@tunnckocore-release-private-key",
-    "WEBHOOK_SECRET": "@tunnckocore-release-webhook-secret"
-  }
-}
